1. 更新依赖组件; 2. 更新数据库结构; 3. 解决Mapper找不到对应xml文件的路径问题; 4. 解决slf4j框架和springboot-logging日志依赖冲突问题; 5. 更新数据库分页器; 6. 修复Swagger2的路由映射报错问题; 7. 更新数据库表结构（主要解决zerofill的问题以及重命名一部分表名）; 8. 解决MyBatis Generator将tinyint(1)直接映射为Byte的问题；9. 完成获取公寓列表、创建公寓、获取员工列表、注册新员工4个接口的开发，并测试通过

+++ src/main/java/com/gbtech/iaaas/mbg/model/AcPublicAccessControl.java
@@ -12,7 +12,7 @@ public class AcPublicAccessControl implements Serializable {
     private Integer acId;
 
     @ApiModelProperty(value = "状态，0: 未启用，1: 已启用")
-    private Byte status;
+    private Integer status;
 
     @ApiModelProperty(value = "创建时间，格式为: 2021-12-31 00:00:00")
     private Date createTime;
@@ -38,11 +38,11 @@ public class AcPublicAccessControl implements Serializable {
         this.acId = acId;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
